% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par10.R
\name{par10}
\alias{par10}
\title{par10 Function}
\usage{
par10(G_file, has_header = FALSE)
}
\arguments{
\item{G_file}{The path to the CSV file containing the matrix of values.}

\item{has_header}{Logical, indicating if the CSV file has a header row. Default is FALSE.}
}
\value{
A matrix where each 0 value is replaced by 10 times the maximum non-zero value in its corresponding row.
}
\description{
This function reads a CSV file containing a matrix of values and modifies it as follows:
If a value in the matrix is 0, it is replaced by 10 times the maximum non-zero value in the corresponding row.
}
\examples{

par10_result <- par10(system.file("extdata", "R.csv", package = "ppRank"), has_header = TRUE)


}
